requirements.txt
pandas
matplotlib
seaborn
scikit-learn
tensorflow
keras
transformers
tf-keras
streamlit-lottie
streamlit
python-dotenv